Externalize CSS and JavaScript code
